+++ plugin-lsp/src/protocol.rs
@@ -702,7 +702,7 @@ impl WorkspaceEdit {
                     }
                     if open_options.open(path).is_err() && !op.ignore_if_exists {
                         editor
-                            .status_bar
+                            .logger
                             .write(MessageKind::Error)
                             .fmt(format_args!("could not create file {:?}", path));
                     }
@@ -719,13 +719,10 @@ impl WorkspaceEdit {
 
                     if op.overwrite || !new_path.exists() || !op.ignore_if_exists {
                         if fs::rename(old_path, new_path).is_err() && !op.ignore_if_exists {
-                            editor
-                                .status_bar
-                                .write(MessageKind::Error)
-                                .fmt(format_args!(
-                                    "could not rename file {:?} to {:?}",
-                                    old_path, new_path
-                                ));
+                            editor.logger.write(MessageKind::Error).fmt(format_args!(
+                                "could not rename file {:?} to {:?}",
+                                old_path, new_path
+                            ));
                         }
                     }
                 }
@@ -738,14 +735,14 @@ impl WorkspaceEdit {
                     if op.recursive {
                         if fs::remove_dir_all(path).is_err() && !op.ignore_if_not_exists {
                             editor
-                                .status_bar
+                                .logger
                                 .write(MessageKind::Error)
                                 .fmt(format_args!("could not delete directory {:?}", path));
                         }
                     } else {
                         if fs::remove_file(path).is_err() && !op.ignore_if_not_exists {
                             editor
-                                .status_bar
+                                .logger
                                 .write(MessageKind::Error)
                                 .fmt(format_args!("could not delete file {:?}", path));
                         }