+++ pepper/rc/changelog.md
@@ -11,6 +11,7 @@
 - added remedybg plugin (under the `plugin-remedybg` folder)
 - added css syntax
 - changed bracket objects to invert bracket positions if invoked with the closing bracket. that is, `a)` will now select text between `)` and `(` instead of `(` and `)`
+- changed `StatusBar` to `Logger` (and `editor.status_bar` to `editor.logger`)
 - added logging to the editor which you can open it with the `open-log` command
 
 # 0.27.0